decided channelinput component does not make any sense
